pyproject.toml
[project]
name = "bazar-market-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boundless-aiogram>=2.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "starlette>=1.0.0",
    "uvicorn>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.15.0",
    "redis>=7.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
]
